requirements.txt
google-generativeai
gTTS
moviepy
google-auth-oauthlib
google-api-python-client
python-dotenv
selenium
